+++ src/gridvoting_jax/models/__init__.py
@@ -1,0 +1,6 @@
+"""Voting models module."""
+
+from .base import VotingModel
+from .spatial import SpatialVotingModel
+
+__all__ = ['VotingModel', 'SpatialVotingModel']